+++ app/src/entities/MapTile.ts
@@ -1,12 +1,10 @@
 import PowerUp from "./PowerUp"
-import {PROP_SEMI_SOLID, TileSetItem, TileSetProperty} from "../types/TileSet"
+import {PROP_SEMI_SOLID, Tile, TileSetProperty} from "../types/TileSet"
 import Player from "./Player"
 import {WorldLayer} from "../types/World"
-import {TileModel} from "../types/TileModel"
+import MapTileModel from "../types/MapTileModel"
 
-export const TILE_SIZE = 64
-
-export default class Tile {
+export default class MapTile {
   id: number
   x: number
   y: number
@@ -25,21 +23,21 @@ export default class Tile {
   power_up: PowerUp | undefined
   properties: TileSetProperty[]
 
-  constructor(id: number, x: number, y: number, layer: WorldLayer, item: TileSetItem) {
+  constructor(id: number, x: number, y: number, layer: WorldLayer, item: Tile) {
     this.id = id
-    this.x = x * TILE_SIZE
-    this.y = y * TILE_SIZE
+    this.x = x * item.tilewidth
+    this.y = y * item.tileheight
     this.name = item.name
     this.source = item.source
     this.offset_x = item.offset_x
     this.offset_y = item.offset_y
+    this.width = item.tilewidth
+    this.height = item.tileheight
     this.tilewidth = item.tilewidth
     this.tileheight = item.tileheight
     this.type = item.type
     this.version = item.version
     this.tiledversion = item.tiledversion
-    this.height = TILE_SIZE
-    this.width = TILE_SIZE
     this.layer = layer.name
     this.power_up = undefined
     this.properties = item.properties || []
@@ -71,9 +69,11 @@ export default class Tile {
     return "/texture/" + hash + ".webp"
   }
 
-  static toModel = (tile: Tile): TileModel => ({
+  static toModel = (tile: MapTile): MapTileModel => ({
     x: tile.x,
     y: tile.y,
+    ox: tile.offset_x,
+    oy: tile.offset_y,
     w: tile.width,
     h: tile.height,
     t: tile.id,